LF-5215 Move chooseIdentification to util file to fix sort.test
